// src/utils/types/StyleType.ts
import { StyleProp} from 'react-native';

export type StyleType = {
    title?: StyleProp<any>;
    titleBold?: StyleProp<any>;
    subtitle?: StyleProp<any>;
    subtitleBold?: StyleProp<any>;
    body?: StyleProp<any>;
    bodyBold?: StyleProp<any>;
    caption?: StyleProp<any>;
    captionBold?: StyleProp<any>;
    medium?: StyleProp<any>;
    mediumBold?: StyleProp<any>;
    small?: StyleProp<any>;
    smallBold?: StyleProp<any>;
    header?: StyleProp<any>;
    footer?: StyleProp<any>;
    primaryButton?: StyleProp<any>;
    secondaryButton?: StyleProp<any>;
    primaryText?: StyleProp<any>;
    secondaryText?: StyleProp<any>;
    lightText?: StyleProp<any>;
    darkText?: StyleProp<any>;
    errorText?: StyleProp<any>;
    successText?: StyleProp<any>;
    warningText?: StyleProp<any>;
    infoText?: StyleProp<any>;
    dangerText?: StyleProp<any>;
    lightBackground?: StyleProp<any>; 
    lightBorder?: StyleProp<any>;
    darkBorder?: StyleProp<any>;
    background?: StyleProp<any>;
}